requirements.txt
asyncio
asyncpg
boto3==1.9.124
discord.py
Django==3.1.7
django-admin-autoregister
django-background-tasks
django-bootstrap-breadcrumbs
django-crispy-forms
django-debug-toolbar-template-timings
django-multiselectfield==0.1.12
django-redis-cache
django-select2==7.2.2
django-sendgrid-v5==0.8.0
django-storages==1.7.1
django-widget-tweaks==1.4.8
djangorestframework
gunicorn
imgurpython==1.1.7
pandas==1.2.3
pinax-messages
psycopg2-binary
pytz==2018.9
redis
requests==2.21.0
whitenoise==5.2.0
numpy
Pillow